+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Avoid trying to issue an `id_token` with client credentials grant
+
 ## [0.2.1] - 2025-10-16
 
 ### Added